requirements.txt
fastapi>=0.104.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
asyncpg>=0.29.0
cohere>=4.27
google-generativeai>=0.3.0
qdrant-client>=1.7.0
httpx>=0.25.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
alembic>=1.13.1
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
tiktoken>=0.5.1